account info
delete will delete all account info associated with it
updating account info will change username. name, and email all at once every time. so some deliniation has to happen. if nothing is in the fields, a blank will be updated into the account
when registering all the values are now saved in various parts of the database